+++ EDFCode/CustomMath.h
@@ -10,8 +10,8 @@
 
 class CustomMath {
   public:
-    static bool ComputeDelta(int16_t _list[], int _size, int16_t *_delta);
-    static bool ComputeMean(int16_t _list[], int _size, int16_t _deltaThreshold, float *_mean);
+    static bool ComputeDelta(float _list[], int _size, int16_t *_delta);
+    static bool ComputeMean(float _list[], int _size, int16_t _deltaThreshold, float *_mean);
     static void VectorNormalize(float _vectorIn[], const int vectorSize);
 };
